refactor: 기존 클래스 패키지 정리
- port-adapter 패턴을 구현하기 위함
