+++ hhplus-tdd-jvm/src/main/kotlin/io/hhplus/tdd/adapter/point/api/PointController.kt
@@ -1,5 +1,7 @@
-package io.hhplus.tdd.point
+package io.hhplus.tdd.adapter.point.api
 
+import io.hhplus.tdd.domain.point.PointHistory
+import io.hhplus.tdd.domain.point.UserPoint
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.web.bind.annotation.*
