+++ hhplus-tdd-jvm/src/main/kotlin/io/hhplus/tdd/domain/point/UserPoint.kt
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point
+package io.hhplus.tdd.domain.point
 
 data class UserPoint(
     val id: Long,
