+++ hhplus-tdd-jvm/src/main/kotlin/io/hhplus/tdd/infrastruture/api/ApiControllerAdvice.kt
@@ -1,4 +1,4 @@
-package io.hhplus.tdd
+package io.hhplus.tdd.infrastruture.api
 
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
