+++ hhplus-tdd-jvm/src/main/kotlin/io/hhplus/tdd/infrastruture/database/PointHistoryTable.kt
@@ -1,7 +1,7 @@
-package io.hhplus.tdd.database
+package io.hhplus.tdd.infrastruture.database
 
-import io.hhplus.tdd.point.PointHistory
-import io.hhplus.tdd.point.TransactionType
+import io.hhplus.tdd.domain.point.PointHistory
+import io.hhplus.tdd.domain.point.TransactionType
 import org.springframework.stereotype.Component
 
 /**
