+++ hhplus-tdd-jvm/src/main/kotlin/io/hhplus/tdd/infrastruture/database/UserPointTable.kt
@@ -1,6 +1,6 @@
-package io.hhplus.tdd.database
+package io.hhplus.tdd.infrastruture.database
 
-import io.hhplus.tdd.point.UserPoint
+import io.hhplus.tdd.domain.point.UserPoint
 import org.springframework.stereotype.Component
 
 /**
